streamline infra contextHolder to other ContextHolders, change getNextAction() to be able to handle CallActions, remove separated Infra Call Progression event and use normal seff progression event instead.
